[css] do not hide download links, use opacity: 0.05
Do not use display:none for download links. This hides them from screen
readers and makes them inaccessible for their users.

The new hide/show method employs use of opacity. In theory, one can use
opacity=0 to render an element completely transparent, but here we use
a slightly greater opacity of 0.05. This gives a visual clue, that the
links are really there, and not missing.

The new method also allows keyboard users to get to the links through
the keyboard.