Allow alphanumeric IDs in source book templates
Example of a source template file for a book:

    {text:-1}
    {text:-2}
    {text:-3}

If for some reason we want to insert a text between 2 and 3, we must regenerate all existing files so that we can insert an integer number inbetween.
This will give us following:

    {text:-1}
    {text:-2}
    {text:-3}
    {text:-4}

As this has no effect on the generated real IDs, it would make sense to allow some other schema for such additional IDs.
For this reason it is now possible to use alphanumeric IDs when such a need arises.
Example:

    {text:-1}
    {text:-2}
    {text:-2a}
    {text:-3}