[SF5 upgrade] Use simple key-value pairs for the download configuration
The parameters in Symfony have only one level, so there does not exist the notion of an array of parameters with a given name.

In order for us to have a simple way to override any of the download configuration parameters,
we need to represent them as simple key-value variables.

This means that we must revert the array logic which was introduced some days ago.
See https://github.com/chitanka/chitanka/commit/19c279b3211eb83aabdad3c26899ccdd9d530e83